requirements.txt
streamlit==1.6.0
pandas==1.3.3
mysql-connector-python==8.0.26
python-dotenv==0.19.2
chromadb==0.0.1
langchain==0.0.1
torch==1.10.0
transformers==4.15.0
